adding authorization policies

+++ app/Http/Controllers/SongbookSongController.php
@@ -16,6 +16,8 @@ class SongbookSongController extends Controller
         ]);
 
         $songbook = Songbook::findOrFail($request->songbook_id);
+        $this->authorize('update', $songbook);
+
         $songbook->songs()->syncWithoutDetaching([$request->song_id]);
 
         return redirect()->back()->with('success');
@@ -23,6 +25,8 @@ class SongbookSongController extends Controller
 
     public function destroy(Songbook $songbook, Song $song)
     {
+        $this->authorize('update', $songbook);
+
         $songbook->songs()->detach($song->id);
         return redirect()->back()->with('success', 'Song removed from songbook.');
     }
